Refactor Utils 2 Electric Boogaloo (#1817)
* Move fhir data evaluation functions to their own service.

* Move some components out of utils

* Fix imports

* fix more imports
